FGDC Geologic Pattern Data

Geologic map and lithology patterns as defined in the FGDC Digital Cartographic Standard for Geologic Map Symbolization by the U.S. Geological Survey and the Geologic Data Subcommittee (GDS) of the Federal Geographic Data Committee (FGDC). deeptime::geo_grob() and deeptime::geo_pattern() should be used to retrieve and modify an individual pattern as a grob object or GridPattern object, respectively.

Usage

geo_grobs

Format

A list, where each item corresponds to a geologic pattern stored as a gTree object as returned by grImport2::pictureGrob(). The names of the list correspond to the pattern codes.

Source

https://ngmdb.usgs.gov/fgdc_gds/geolsymstd.php via https://github.com/davenquinn/geologic-patterns

Details

For specific pattern codes, see the "pattern numbers" in the full pattern chart. Daven Quinn has also assembled more accessible documentation of the map patterns/codes and lithology patterns/codes. rmacrostrat::def_lithologies() can also be used to look up pattern codes for various lithologies (see the "fill" column). Note that patterns associated with color variants (e.g., "101-M") are not included but can be created using deeptime::geo_grob().

These patterns were originally processed and optimized by Daven Quinn and are hosted on GitHub.